pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "microscopy-ai-agent-demo"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "smolagents[toolkit,mcp,litellm,transformers]",
    "litellm==1.82.2",
    "bitsandbytes",
    "psutil",
    "wikipedia-api",
    "twisted",
    "numpy",
    "Pyro5",
    "numba",
    "SciFiReaders",
    "sidpy",
    "scipy",
    "matplotlib",
    "scikit-image",
    "pydantic",
    "pydantic-settings",
    "pytest",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "pyyaml",
    "graphviz",
    "pydot",
    "pandas",
    "Pillow",
    "h5py",
    "torch",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["atomonous*"]

[tool.uv]
override-dependencies = [
    "litellm==1.82.2",
]
